+++ CHANGES.md
@@ -1,10 +1,12 @@
-## Changelog
+# Changelog
 
 ### 1.9.0
+
 * All: Update `make test` runner to use `pytest` instead of deprecated `setup.py test`
 * pytest-server-fixtures: Update deprecated mongo 7.* to 8.2
 * pytest-profile: add support for `cProfile` regex to filter output
 * pytest-profile: update README.md args
+* pytest-virtualenv: fix: Install local dir to virtualenv in later py3
 * ci: fix py3.7-3.13 ubuntu
 * ci: Update support in tests for py3.7-13 for ubuntu 22.04